Migrate test/src/common to null safety
