+++ packages/flutter_tools/test/commands.shard/hermetic/analyze_continuously_test.dart
@@ -24,6 +24,7 @@ import 'package:process/process.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fake_process_manager.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   setUpAll(() {
