+++ packages/flutter_tools/test/commands.shard/hermetic/build_fuchsia_test.dart
@@ -22,6 +22,7 @@ import 'package:test/fake.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 // Defined globally for fakes to use.
 FileSystem fileSystem;
