+++ packages/flutter_tools/test/commands.shard/hermetic/build_linux_test.dart
@@ -23,6 +23,7 @@ import 'package:test/fake.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 const String _kTestFlutterRoot = '/flutter';
