+++ packages/flutter_tools/test/commands.shard/hermetic/build_test.dart
@@ -10,6 +10,7 @@ import 'package:flutter_tools/src/runner/flutter_command.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   testUsingContext('obfuscate requires split-debug-info', () {
