+++ packages/flutter_tools/test/commands.shard/hermetic/build_web_test.dart
@@ -22,6 +22,7 @@ import 'package:flutter_tools/src/web/compile.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   FileSystem fileSystem;
