+++ packages/flutter_tools/test/commands.shard/hermetic/config_test.dart
@@ -20,6 +20,7 @@ import 'package:test/fake.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   FakeAndroidStudio fakeAndroidStudio;
