+++ packages/flutter_tools/test/commands.shard/hermetic/devices_test.dart
@@ -16,6 +16,7 @@ import 'package:flutter_tools/src/globals.dart' as globals;
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fake_devices.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   group('devices', () {
