+++ packages/flutter_tools/test/commands.shard/hermetic/downgrade_test.dart
@@ -17,6 +17,7 @@ import 'package:mockito/mockito.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   FileSystem fileSystem;
