+++ packages/flutter_tools/test/commands.shard/hermetic/drive_test.dart
@@ -17,6 +17,7 @@ import 'package:test/fake.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   FileSystem fileSystem;
