+++ packages/flutter_tools/test/commands.shard/hermetic/precache_test.dart
@@ -13,6 +13,7 @@ import 'package:mockito/mockito.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   MockCache cache;
