+++ packages/flutter_tools/test/commands.shard/hermetic/pub_get_test.dart
@@ -16,6 +16,7 @@ import 'package:test/fake.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   FileSystem fileSystem;
