+++ packages/flutter_tools/test/commands.shard/hermetic/run_test.dart
@@ -32,6 +32,7 @@ import 'package:vm_service/vm_service.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   group('run', () {
