+++ packages/flutter_tools/test/commands.shard/hermetic/symbolize_test.dart
@@ -21,6 +21,7 @@ import 'package:test/fake.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   MemoryFileSystem fileSystem;
