+++ packages/flutter_tools/test/commands.shard/hermetic/test_test.dart
@@ -25,6 +25,7 @@ import 'package:process/process.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fake_devices.dart';
+import '../../src/test_flutter_command_runner.dart';
 import '../../src/testbed.dart';
 
 const String _pubspecContents = '''
