+++ packages/flutter_tools/test/commands.shard/permeable/build_apk_test.dart
@@ -23,6 +23,7 @@ import '../../src/android_common.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/mocks.dart' hide MockAndroidSdk;
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   Cache.disableLocking();
