+++ packages/flutter_tools/test/commands.shard/permeable/build_appbundle_test.dart
@@ -18,6 +18,7 @@ import 'package:test/fake.dart';
 import '../../src/android_common.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   Cache.disableLocking();
