+++ packages/flutter_tools/test/general.shard/analytics_test.dart
@@ -27,6 +27,7 @@ import 'package:usage/usage_io.dart';
 import '../src/common.dart';
 import '../src/context.dart';
 import '../src/fakes.dart';
+import '../src/test_flutter_command_runner.dart';
 
 void main() {
   setUpAll(() {
