+++ packages/flutter_tools/test/general.shard/base/logs_test.dart
@@ -10,6 +10,7 @@ import 'package:flutter_tools/src/commands/logs.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/test_flutter_command_runner.dart';
 
 void main() {
   group('logs', () {
