+++ packages/flutter_tools/test/general.shard/channel_test.dart
@@ -14,6 +14,7 @@ import 'package:flutter_tools/src/version.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
+import '../src/test_flutter_command_runner.dart';
 
 void main() {
   group('channel', () {
