+++ packages/flutter_tools/test/general.shard/drive/drive_service_test.dart
@@ -23,6 +23,7 @@ import 'package:vm_service/vm_service.dart' as vm_service;
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/fake_vm_services.dart';
 import '../../src/fakes.dart';
