+++ packages/flutter_tools/test/general.shard/flutter_validator_test.dart
@@ -18,6 +18,18 @@ import '../src/common.dart';
 import '../src/context.dart';
 import '../src/fakes.dart';
 
+/// Matches a doctor validation result.
+Matcher _matchDoctorValidation({
+  ValidationType validationType,
+  String statusInfo,
+  dynamic messages
+}) {
+  return const TypeMatcher<ValidationResult>()
+      .having((ValidationResult result) => result.type, 'type', validationType)
+      .having((ValidationResult result) => result.statusInfo, 'statusInfo', statusInfo)
+      .having((ValidationResult result) => result.messages, 'messages', messages);
+}
+
 void main() {
   testWithoutContext('FlutterValidator shows an error message if gen_snapshot is '
     'downloaded and exits with code 1', () async {
@@ -48,7 +60,7 @@ void main() {
     fileSystem.file(artifacts.getArtifactPath(Artifact.genSnapshot)).createSync(recursive: true);
 
 
-    expect(await flutterValidator.validate(), matchDoctorValidation(
+    expect(await flutterValidator.validate(), _matchDoctorValidation(
       validationType: ValidationType.partial,
       statusInfo: 'Channel unknown, 1.0.0, on Linux, locale en_US.UTF-8',
       messages: containsAll(const <ValidationMessage>[
@@ -84,7 +96,7 @@ void main() {
 
     // gen_snapshot is downloaded on demand, and the doctor should not
     // fail if the gen_snapshot binary is not present.
-    expect(await flutterValidator.validate(), matchDoctorValidation(
+    expect(await flutterValidator.validate(), _matchDoctorValidation(
       validationType: ValidationType.installed,
       statusInfo: 'Channel unknown, 1.0.0, on Windows, locale en_US.UTF-8',
       messages: anything,
@@ -103,7 +115,7 @@ void main() {
       flutterRoot: () => 'sdk/flutter',
     );
 
-    expect(await flutterValidator.validate(), matchDoctorValidation(
+    expect(await flutterValidator.validate(), _matchDoctorValidation(
       validationType: ValidationType.partial,
       statusInfo: 'Channel unknown, 0.0.0, on Windows, locale en_US.UTF-8',
       messages: const <ValidationMessage>[
@@ -138,7 +150,7 @@ void main() {
       flutterRoot: () => 'sdk/flutter'
     );
 
-    expect(await flutterValidator.validate(), matchDoctorValidation(
+    expect(await flutterValidator.validate(), _matchDoctorValidation(
       validationType: ValidationType.installed,
       statusInfo: 'Channel unknown, 1.0.0, on Windows, locale en_US.UTF-8',
       messages: containsAll(const <ValidationMessage>[
