+++ packages/flutter_tools/test/general.shard/fuchsia/fuchsia_device_test.dart
@@ -41,6 +41,7 @@ import 'package:vm_service/vm_service.dart' as vm_service;
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/fake_vm_services.dart';
 
 final vm_service.Isolate fakeIsolate = vm_service.Isolate(
   id: '1',
