+++ packages/flutter_tools/test/general.shard/integration_test_device_test.dart
@@ -16,6 +16,7 @@ import 'package:vm_service/vm_service.dart' as vm_service;
 import '../src/common.dart';
 import '../src/context.dart';
 import '../src/fake_devices.dart';
+import '../src/fake_vm_services.dart';
 
 final Map<String, Object> vm = <String, dynamic>{
   'isolates': <dynamic>[
