+++ packages/flutter_tools/test/general.shard/ios/ios_device_logger_test.dart
@@ -21,6 +21,7 @@ import 'package:vm_service/vm_service.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/fake_vm_services.dart';
 
 void main() {
   FakeProcessManager processManager;
