+++ packages/flutter_tools/test/general.shard/resident_runner_test.dart
@@ -40,6 +40,7 @@ import 'package:mockito/mockito.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
+import '../src/fake_vm_services.dart';
 import '../src/fakes.dart';
 import '../src/testbed.dart';
