+++ packages/flutter_tools/test/general.shard/runner/flutter_command_test.dart
@@ -25,6 +25,7 @@ import 'package:mockito/mockito.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 import '../../src/fakes.dart';
+import '../../src/test_flutter_command_runner.dart';
 import 'utils.dart';
 
 void main() {
